the structure of bot's modes was refactored
